removed dictionary and multi dictionary; fixed 'encoding::unhex()'

+++ src/net/request_context.h
@@ -25,8 +25,8 @@ __NET_BEGIN__
 
 struct ProtocolVersion
 {
-	size_t major{};
-	size_t minor{};
+	unsigned short major{};
+	unsigned short minor{};
 
 	inline bool operator== (const ProtocolVersion& other) const
 	{
